backend/requirements.txt
fastapi~=0.115.6
uvicorn~=0.22.0
sqlalchemy~=1.4.46
requests~=2.32.3
psycopg2-binary
python-dotenv
pydantic~=1.10.7
PyJWT~=2.10.1
starlette~=0.41.3
python-jose[cryptography]
pycryptodome
yfinance
python-dotenv
pandas
cachetools
ta